Move canonical ABI function trait aliases
Place it in the `traits` module, in preparation to create a new
`InstanceWithMemory` trait alias to be placed beside
`InstanceWithFunction`.
